test: move the maindispatcherrule to the testing module
The CoroutinesMainDispatcherRule will be used across every module, so move it into the testing
module to be shared
