+++ app/src/test/java/com/chesire/malime/flow/login/details/DetailsViewModelTests.kt
@@ -2,7 +2,7 @@ package com.chesire.malime.flow.login.details
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
-import com.chesire.malime.CoroutinesMainDispatcherRule
+import com.chesire.malime.testing.CoroutinesMainDispatcherRule
 import com.chesire.malime.account.UserRepository
 import com.chesire.malime.core.models.UserModel
 import com.chesire.malime.server.Resource
