+++ app/src/test/java/com/chesire/malime/flow/series/detail/SeriesDetailViewModelTests.kt
@@ -3,7 +3,7 @@ package com.chesire.malime.flow.series.detail
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
 import com.chesire.malime.AuthCaster
-import com.chesire.malime.CoroutinesMainDispatcherRule
+import com.chesire.malime.testing.CoroutinesMainDispatcherRule
 import com.chesire.malime.core.flags.AsyncState
 import com.chesire.malime.series.SeriesRepository
 import com.chesire.malime.server.Resource
