+++ testing/build.gradle
@@ -28,4 +28,6 @@ dependencies {
     implementation project(path: ':core')
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:$coroutines_version"
+    implementation 'junit:junit:4.12'
 }
